CORE-249: Address “value stored in `var` is never read”
Add a macro ANALYZER_IGNORE_UNREAD_VARIABLE, that only applies when __clang_analyzer__ is defined, to silence the analyzer warning about unread variables.  Current code keeps unread variables around as documentation and should not be removed - so we explicitly tell the analyzer that the report is not an issue.  Added support/BRSupport.h