Remove balance update event on insert wallet
At this stage, the wallet balance will always be 0, which is the default value, so I think it is too early to be sending a WALLET_EVENT_BALANCE_UPDATED event.

The event really means “We know what the balance is now” to the receiver, so I think we should only send the event when we get a result from an updateBalance call.